+++ git-packetline/src/line/blocking_io.rs
@@ -1,32 +1,28 @@
 use std::io;
 
-use crate::{
-    encode,
-    immutable::{Band, Error, Text},
-    Channel, PacketLineRef,
-};
+use crate::{encode, BandRef, Channel, ErrorRef, PacketLineRef, TextRef};
 
-impl<'a> Band<'a> {
+impl<'a> BandRef<'a> {
     /// Serialize this instance to `out`, returning the amount of bytes written.
     ///
     /// The data written to `out` can be decoded with [`Borrowed::decode_band()]`.
     pub fn write_to(&self, out: impl io::Write) -> io::Result<usize> {
         match self {
-            Band::Data(d) => encode::band_to_write(Channel::Data, d, out),
-            Band::Progress(d) => encode::band_to_write(Channel::Progress, d, out),
-            Band::Error(d) => encode::band_to_write(Channel::Error, d, out),
+            BandRef::Data(d) => encode::band_to_write(Channel::Data, d, out),
+            BandRef::Progress(d) => encode::band_to_write(Channel::Progress, d, out),
+            BandRef::Error(d) => encode::band_to_write(Channel::Error, d, out),
         }
     }
 }
 
-impl<'a> Text<'a> {
+impl<'a> TextRef<'a> {
     /// Serialize this instance to `out`, appending a newline if there is none, returning the amount of bytes written.
     pub fn write_to(&self, out: impl io::Write) -> io::Result<usize> {
         encode::text_to_write(self.0, out)
     }
 }
 
-impl<'a> Error<'a> {
+impl<'a> ErrorRef<'a> {
     /// Serialize this line as error to `out`.
     ///
     /// This includes a marker to allow decoding it outside of a side-band channel, returning the amount of bytes written.