+++ git-packetline/src/read/sidebands/async_io.rs
@@ -7,11 +7,7 @@ use std::{
 use futures_io::{AsyncBufRead, AsyncRead};
 use futures_lite::ready;
 
-use crate::{
-    decode,
-    immutable::{Band, Text},
-    PacketLineRef, StreamingPeekableIter, U16_HEX_BYTES,
-};
+use crate::{decode, BandRef, PacketLineRef, StreamingPeekableIter, TextRef, U16_HEX_BYTES};
 
 type ReadLineResult<'a> = Option<std::io::Result<Result<PacketLineRef<'a>, decode::Error>>>;
 /// An implementor of [`AsyncBufRead`] yielding packet lines on each call to [`read_line()`][AsyncBufRead::read_line()].
@@ -238,13 +234,13 @@ where
                                         .map_err(|err| io::Error::new(io::ErrorKind::Other, err))?;
                                     const ENCODED_BAND: usize = 1;
                                     match band {
-                                        Band::Data(d) => break (U16_HEX_BYTES + ENCODED_BAND, d.len()),
-                                        Band::Progress(d) => {
-                                            let text = Text::from(d).0;
+                                        BandRef::Data(d) => break (U16_HEX_BYTES + ENCODED_BAND, d.len()),
+                                        BandRef::Progress(d) => {
+                                            let text = TextRef::from(d).0;
                                             handle_progress(false, text);
                                         }
-                                        Band::Error(d) => {
-                                            let text = Text::from(d).0;
+                                        BandRef::Error(d) => {
+                                            let text = TextRef::from(d).0;
                                             handle_progress(true, text);
                                         }
                                     };