+++ git-packetline/CHANGELOG.md
@@ -4,5 +4,8 @@
 
 * **renames / moves**
     - `immutable::PacketLine` -> `PacketLineRef`
+    - `immutable::Error` -> `ErrorRef`
+    - `immutable::Text` -> `TextRef`
+    - `immutable::Band` -> `BandRef`
     - `immutable::DecodeBandError` -> `decode::band::Error`
-
+    - `pub immutable::` -> `line::`